requirements.txt
dlib
face-recognition
opencv-python
numpy
pillow
flask
